Refactoring logins overview
