+++ src/pages/account/logins/index.tsx
@@ -6,8 +6,8 @@ import type { SortedIdentities } from "@src/@types/sorted-identities";
 import { UserCookie } from "@src/@types/user-cookie";
 import { NoDataReturned, NoIdentityFound } from "@src/client-errors";
 import { Container } from "@src/components/containers/container";
-import { LoginsOverview } from "@src/components/logins-overview/logins-overview";
 import { Layout } from "@src/components/page-layout";
+import { LoginsOverview } from "@src/components/pages/logins-overview/logins-overview";
 import { GraphqlErrors } from "@src/errors";
 import { withAuth } from "@src/middlewares/with-auth";
 import { withLogger } from "@src/middlewares/with-logger";
